Fixing dunit build and nearly all failures
 * adjusting heap-sizes, env/system properties, and excluding WAN tests from distributedTest target in build.gradle
 * fixing failures in ConcurrencyChecksDUnit in parallel runs by waiting for async KRF creation to complete before writing the index idxkrf file;
   other changes to avoid unnecessary File*Stream creation, using new Files API to get proper exception etc
 * fixing failures in reconnect tests by assuming LogWriterImpl but not PureLogWriter for standard system logger
 * fixing failures in native timer tests by avoiding NanoTimer initialization in static PartitionedRegion initializations
 * don't create persistent PDX_types region for non-WAN cases to fix some failures in AsyncQueue tests
 * fixing occasional failures in HA tests by adding failover for DATABASE create/boot failures
 * correcting exception handling in TransactionResourceImpl by avoiding unwrap when SQLState is null
   (e.g. from custom user procedure code); fixes failures in some DAP tests
 * fixing failure in BackwardCompatabilityDUnit.testDiskStoreUpgradeAndClientCompatibility_SQLF_or_GFXD_Versions_To_CurrentGemFireXD
   and disabling all others since we won't be supporting rolling upgrade
 * updating GemFireXDAuthenticationDUnit to use embedded LDAP server (Apache Directory Server) instance
   like used by new LdapGroupDUnit tests which is started once at the start of class and
   closed at end of class (beforeClass/afterClass)
 * updating to use the newer fixed implementation of AnyCyclicBarrier and adding its cleanup in TransactionDUnit
 * moving a couple of class AsyncListener and Price which are used by various DUnit tests
 * making off-heap memory as constant at 500M for all tests and changing test expectations accordingly;
   this is because this setting is a static in a JVM and is done only once by GemFire layer in the entire test run
 * fixing multiple failures in ClientServerDUnit
 * enabled BugsDUnit.DISABLED_BUG_51746_testBug51039 which passes with current code
 * add dropping of HDFS stores in common DistributedSQLTestBase shutdown to avoid cascading failures
 * disabled tests in SingleHopTransactionDUnit (GEMXD-12), GemFireXDAuthenticationDUnit
     .testSchemaSharedBetweenUsersWithoutAuthorization (GEMXD-11), Oracle tests in SerialDBSynchronizerPart*DUnit
 * keeping only p2p.discoveryTimeout and p2p.joinTimeout settings for faster startup in tests but disabling others
   to fix occasional cascading failures due to member being forced out of DS
 * fixing file paths etc in all tests like *HDFS*DUnit, *DBSynchronizer*DUnit, BugsDUnit, CustomEvictionDUnit etc
   to fix multiple failures; also fixing directory cleanup code in many tests
 * fixing multiple hangs due to different reasons (e.g. in connection creation in BugsDUnit)
 * fixing failures in LangScripts_StringTypesDUnit, LangScripts_LOBDUnit, LangScripts_IntersectDUnit,
   LangScripts_ImplicitConversionsDUnit due to CLOB/LONGVARCHAR columns now allowing ordering
   and comparison operations and enhanced some of those tests as per equivalent JUnit tests
 * fixing failures in GfxdDiagsDUnit to allow for either IP address or host name in expected URLs
 * adding waiting for pending commits in many transaction tests to fix occasional failures
 * many other fixes like in HDFSRegionManagementDUnit, DistributedQueryDUnit, InternalManagementServiceDUnit,
   QueryTimeOutDUnit, TransactionDUnit
 * fixing build warnings with Java 8 in ParameterMappingTest, JDBCDriverTest, CastingTest

Still failing: couple of failures in TransactionRRHDFSTableDUnit,
    occasional assertion failure errors in StatementPlan*DUnit

+++ gemfirexd/tools/src/dunit/java/com/pivotal/gemfirexd/DistributedSQLTestBase.java
@@ -42,6 +42,7 @@ import com.gemstone.gemfire.cache.DiskAccessException;
 import com.gemstone.gemfire.cache.Region;
 import com.gemstone.gemfire.cache.RegionAttributes;
 import com.gemstone.gemfire.cache.asyncqueue.AsyncEventQueue;
+import com.gemstone.gemfire.cache.hdfs.internal.HDFSStoreImpl;
 import com.gemstone.gemfire.cache.wan.GatewayReceiver;
 import com.gemstone.gemfire.cache.wan.GatewaySender;
 import com.gemstone.gemfire.distributed.DistributedMember;
@@ -89,7 +90,6 @@ import org.apache.derby.iapi.services.monitor.Monitor;
 import org.apache.derby.iapi.services.stream.HeaderPrintWriter;
 import org.apache.derbyTesting.junit.CleanDatabaseTestSetup;
 import org.apache.derbyTesting.junit.TestConfiguration;
-import org.apache.log4j.Logger;
 import org.junit.internal.MethodSorter;
 import org.w3c.dom.Element;
 import org.xml.sax.SAXException;
@@ -279,23 +279,22 @@ public class DistributedSQLTestBase extends DistributedTestBase {
     }
   }
 
+  public static void startCommonLocator(int locatorPort) {
+    FabricLocator loc = FabricServiceManager.getFabricLocatorInstance();
+    if (loc.status() != FabricService.State.RUNNING) {
+      try {
+        loc.start("localhost", locatorPort, null);
+      } catch (SQLException e) {
+        throw new TestException("Failed to start locator", e);
+      }
+    }
+    assert (loc.status() == FabricService.State.RUNNING);
+  }
+
   public void beforeClass() throws Exception {
     // Start a locator once for whole suite
-    final int locatorPort = getDUnitLocatorPort();
-    DistributedTestBase.invokeInLocator(new SerializableRunnable() {
-      @Override
-      public void run() {
-        FabricLocator loc = FabricServiceManager.getFabricLocatorInstance();
-        if (loc.status() != FabricService.State.RUNNING) {
-          try {
-            loc.start("localhost", locatorPort, null);
-          } catch (SQLException e) {
-            throw new TestException("Failed to start locator", e);
-          }
-        }
-        assert (loc.status() == FabricService.State.RUNNING);
-      }
-    });
+    Host.getLocator().invoke(DistributedSQLTestBase.class,
+        "startCommonLocator", getDUnitLocatorPort());
   }
 
   public void afterClass() throws Exception {
@@ -440,7 +439,7 @@ public class DistributedSQLTestBase extends DistributedTestBase {
     }
   }
 
-  protected String getSysDirName() {
+  protected static String getSysDirName() {
     return new File(".").getAbsolutePath();
   }
 
@@ -538,15 +537,17 @@ public class DistributedSQLTestBase extends DistributedTestBase {
       setGFXDProperty(props, "log-level", logLevel);
     }
     // reduce timeout properties for faster dunit runs
-    setGFXDProperty(props, "member-timeout", "2000");
     System.setProperty("p2p.discoveryTimeout", "1000");
     System.setProperty("p2p.joinTimeout", "2000");
+    /*
+    setGFXDProperty(props, "member-timeout", "2000");
     System.setProperty("p2p.leaveTimeout", "1000");
     System.setProperty("p2p.socket_timeout", "4000");
     System.setProperty("p2p.disconnectDelay", "500");
     System.setProperty("p2p.handshakeTimeoutMs", "2000");
     System.setProperty("p2p.lingerTime", "500");
     System.setProperty("p2p.listenerCloseTimeout", "4000");
+    */
     if (extraProps != null) {
       Enumeration<?> e = extraProps.propertyNames();
       while (e.hasMoreElements()) {
@@ -845,10 +846,10 @@ public class DistributedSQLTestBase extends DistributedTestBase {
   }
 
   public static DistributedMember _startNewServer(String className,
-      String name, int mcastPort, String serverGroups, Properties extraProps, boolean configureOffHeap)
-      throws Exception {
+      String name, int mcastPort, String serverGroups, Properties extraProps,
+      boolean configureOffHeap) throws Exception {
     final Class<?> c = Class.forName(className);
-   
+
     final DistributedSQLTestBase test = (DistributedSQLTestBase)c
         .getConstructor(String.class).newInstance(name);
     test.configureDefaultOffHeap = configureOffHeap;
@@ -2052,6 +2053,10 @@ public class DistributedSQLTestBase extends DistributedTestBase {
             executeCleanup(stmt, "drop gatewayreceiver \"" + receiver.getId()
                 + '"');
           }
+          for (HDFSStoreImpl hdfsStore : cache.getAllHDFSStores()) {
+            executeCleanup(stmt, "drop hdfsstore \"" +
+                hdfsStore.getName() + '"');
+          }
           for (DiskStoreImpl ds : cache.listDiskStores()) {
             if (!GfxdConstants.GFXD_DEFAULT_DISKSTORE_NAME.equals(ds
                 .getName()) && !GfxdConstants.GFXD_DD_DISKSTORE_NAME
@@ -2388,9 +2393,10 @@ public class DistributedSQLTestBase extends DistributedTestBase {
     try {
       deleteDataDictionaryDir();
       deleteGlobalIndexCachinDir();
-      String[] testSpecificDirs = testInstance
-          .testSpecificDirectoriesForDeletion();
-      if (testSpecificDirs != null) {
+      final DistributedSQLTestBase test = testInstance;
+      final String[] testSpecificDirs;
+      if (test != null && (testSpecificDirs = test
+          .testSpecificDirectoriesForDeletion()) != null) {
         for (String file : testSpecificDirs) {
           File dir = new File(getSysDiskDir(), file);
           boolean result = TestUtil.deleteDir(dir);
@@ -2402,6 +2408,28 @@ public class DistributedSQLTestBase extends DistributedTestBase {
       }
     } catch (Exception ex) {
       throw new TestException("unexpected exception", ex);
+    }
+  }
+
+  @SuppressWarnings("ResultOfMethodCallIgnored")
+  public static void delete(File file) {
+    if (!file.exists()) {
+      return;
+    }
+    if (file.isDirectory()) {
+      if (file.list().length == 0) {
+        file.delete();
+      } else {
+        File[] files = file.listFiles();
+        if (files != null) {
+          for (File f : files) {
+            delete(f);
+          }
+        }
+        file.delete();
+      }
+    } else {
+      file.delete();
     }
   }
 
@@ -2560,7 +2588,7 @@ public class DistributedSQLTestBase extends DistributedTestBase {
 
   public static String getSysDiskDir() {
     try {
-      return testInstance.getSysDirName();
+      return getSysDirName();
     } catch (Exception ex) {
       throw new TestException("unexpected exception", ex);
     }
